% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_build_task.R
\name{apprs_build_task}
\alias{apprs_build_task}
\title{Build a task request}
\usage{
apprs_build_task(df, roi)
}
\arguments{
\item{df}{a data frame with task name (task), subtask name (subtask),
latitude (lat), longitude (lon), start (start) and end (end) date.}

\item{roi}{a region of interest defined by a SpatRaster or sf object,
the roi will override any point based data provided as lat/lon coordinates
in the data frame}
}
\value{
a valid AppEEARS JSON formatted task
}
\description{
Build a task request
}
