% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_bundle.R
\name{apprs_bundle}
\alias{apprs_bundle}
\title{AppEEARS list of bundled files to download}
\usage{
apprs_bundle(user, task_id)
}
\arguments{
\item{user}{username used to sign up}
}
\value{
returns a nested list of files to download
}
\description{
Returns a data frame of all data ready for download as one bundle
}
\examples{

\dontrun{
# get a list of datasets
apprs_bundle(
user = "your_user_name",
task_id = "a_task_id"
)
}
}
\author{
Koen Hufkens
}
