% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_delete.R
\name{apprs_delete}
\alias{apprs_delete}
\title{Delete AppEEARS task from queue}
\usage{
apprs_delete(user, task_id, purge = FALSE)
}
\arguments{
\item{user}{username used to sign up}

\item{task_id}{AppEEARS task id}

\item{purge}{if TRUE, remove all previously finished tasks from
the task list (default = FALSE)}
}
\value{
returns the content of the API call
}
\description{
Removes a task from the queue and or buffer
}
\examples{

\dontrun{
# delete a single task
apprs_delete(
  user = "your_user_name",
  task_id = "a_task_id"
)

# delete all finished or crashed
# jobs (if not deleted previously)
apprs_delete(
  user = "your_user_name",
  purge = TRUE
)
}
}
\author{
Koen Hufkens
}
