% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_get_key.R
\name{apprs_get_key}
\alias{apprs_get_key}
\title{Store username and password}
\usage{
apprs_get_key(user)
}
\arguments{
\item{user}{username used to sign up}
}
\value{
the password set using \code{\link[apprs]{apprs_set_key}} saved
in the keychain
}
\description{
Returns you token set by \code{\link[apprs]{apprs_set_key}}
}
\examples{

\dontrun{
# set key
apprs_set_key(user = "test@mail.com", password = "123")

# get key
apprs_get_key(user = "test@mail.com")
}
}
\seealso{
\code{\link[apprs]{apprs_set_key}}
}
\author{
Koen Hufkens
}
