% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_layers.R
\name{apprs_layers}
\alias{apprs_layers}
\title{AppEEARS dataset layers}
\usage{
apprs_layers(product)
}
\arguments{
\item{product}{product for which to list the layers}
}
\value{
returns a data frame with the AppEEARS datasets
}
\description{
Returns a data frame of available layers
for an AppEEARS product
}
\examples{

\dontrun{
# get a list of datasets
apprs_layers()
}
}
\seealso{
\code{\link[apprs]{apprs_products}}
}
\author{
Koen Hufkens
}
