% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_login.R
\name{apprs_logout}
\alias{apprs_logout}
\title{Invalidates an AppEEARS bearer token}
\usage{
apprs_logout(token)
}
\arguments{
\item{token}{a Bearer token as returned by apprs_login()}
}
\description{
Given a token it will log out / delete this
token, invalidating it.
}
