% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_products.R
\name{apprs_products}
\alias{apprs_products}
\title{AppEEARS dataset list}
\usage{
apprs_products()
}
\value{
returns a data frame with the AppEEARS datasets
}
\description{
Returns a data frame of available data products
}
\examples{

\dontrun{
# get a list of datasets
apprs_products()
}
}
\seealso{
\code{\link[apprs]{apprs_set_key}}
\code{\link[apprs]{apprs_transfer}}
\code{\link[apprs]{apprs_request}}
}
\author{
Koen Hufkens
}
