% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_set_key.R
\name{apprs_set_key}
\alias{apprs_set_key}
\title{Set AppEEARS password}
\usage{
apprs_set_key(user, password)
}
\arguments{
\item{user}{user used to sign up for the AppEEARS data service (this is
not the email address, but the true user name!)}

\item{password}{used to sign up for AppEEARS}
}
\value{
It invisibly returns the user.
}
\description{
Saves the token to your local keychain under
a service called "appeears".
}
\details{
In systems without keychain management set the option
keyring_backend to `file` (i.e. options(keyring_backend = "file"))
in order to write the keychain entry to an encrypted file.
This mostly pertains to headless Linux systems. The keychain files
can be found in ~/.config/r-keyring.
}
\examples{

\dontrun{
# set key
apprs_set_key(user = "test", password = "123")

# get key
apprs_get_key(user = "test")

# leave user and key empty to open a browser window to the service's website
# and type the key interactively
apprs_get_key()

}
}
\seealso{
\code{\link[apprs]{apprs_get_key}}
}
\author{
Koen Hufkens
}
