% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_delete.R
\name{rs_delete}
\alias{rs_delete}
\title{Delete AppEEARS task from queue}
\usage{
rs_delete(task_id, user, purge = FALSE)
}
\arguments{
\item{task_id}{AppEEARS task id}

\item{user}{username used to sign up}

\item{purge}{if TRUE, remove all previously finished tasks from
the task list (default = FALSE)}
}
\value{
returns the content of the API call
}
\description{
Removes a task from the queue and or buffer
}
\examples{

\dontrun{
# delete a single task
rs_delete(
  user = "your_user_name",
  task_id = "a_task_id"
)

# delete all finished or crashed
# jobs (if not deleted previously)
rs_delete(
  user = "your_user_name",
  purge = TRUE
)
}
}
\author{
Koen Hufkens
}
