% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_get_key.R
\name{rs_get_key}
\alias{rs_get_key}
\title{Retrieve NASA Earth Data password}
\usage{
rs_get_key(user)
}
\arguments{
\item{user}{username used to sign up}
}
\value{
the password set using \code{\link[appeears]{rs_set_key}} saved
in the keychain
}
\description{
Returns you token set by \code{\link[appeears]{rs_set_key}}
}
\examples{

\dontrun{
# set key
rs_set_key(user = "test@mail.com", password = "123")

# get key
rs_get_key(user = "test@mail.com")
}
}
\seealso{
\code{\link[appeears]{rs_set_key}}
}
\author{
Koen Hufkens
}
