% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_layers.R
\name{rs_layers}
\alias{rs_layers}
\title{AppEEARS dataset layers}
\usage{
rs_layers(product)
}
\arguments{
\item{product}{product for which to list the layers}
}
\value{
returns a data frame with the AppEEARS datasets
}
\description{
Returns a data frame of available layers
for an AppEEARS product
}
\examples{

\dontrun{
# get a list of datasets
rs_layers()
}
}
\seealso{
\code{\link[appeears]{rs_products}}
}
\author{
Koen Hufkens
}
