% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_login.R
\name{rs_login}
\alias{rs_login}
\title{Checks AppEEARS login}
\usage{
rs_login(user)
}
\arguments{
\item{user}{AppEEARS username}
}
\value{
returns an AppEEARS session (bearer) token
}
\description{
Returns a valid token for a session if
successful otherwise fails with an error (stop())
}
