% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_login.R
\name{rs_logout}
\alias{rs_logout}
\title{Invalidates an AppEEARS bearer token}
\usage{
rs_logout(token)
}
\arguments{
\item{token}{a Bearer token as returned by rs_login()}
}
\value{
returns if the session has closed TRUE/FALSE
}
\description{
Given a token it will log out / delete this
token, invalidating it.
}
