% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_products.R
\name{rs_products}
\alias{rs_products}
\title{AppEEARS dataset list}
\usage{
rs_products()
}
\value{
returns a data frame with the AppEEARS datasets
}
\description{
Returns a data frame of available data products
}
\examples{

# is the server reachable
server_check <- appeears:::rs_running(
 file.path(appeears:::rs_server(),"product")
)

# get a list of datasets
if(server_check){
 products <- rs_products()
}
}
\seealso{
\code{\link[appeears]{rs_set_key}}
\code{\link[appeears]{rs_transfer}}
\code{\link[appeears]{rs_request}}
}
\author{
Koen Hufkens
}
