% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_quality.R
\name{rs_quality}
\alias{rs_quality}
\title{List or translate AppEEARS quality metrics}
\usage{
rs_quality(product, layer, value)
}
\arguments{
\item{product}{AppEEARS product name}

\item{layer}{name of a product quality control layer}

\item{value}{quality control value to translate}
}
\value{
returns a data frame of all AppEEARS quality layers,
or those associated with a product. When a value is provided
this quality flag will be translated from bitwise representation
to plain language.
}
\description{
Returns a data frame of all quality layers,
or the translation of a quality layer value
into plain language.
}
\examples{

\dontrun{
# get a list of quality layers for all data products
rs_quality()
}
}
\author{
Koen Hufkens
}
