% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_request.R
\name{apprs_request}
\alias{apprs_request}
\title{AppEEARS data request and download}
\usage{
apprs_request(
  request,
  user,
  transfer = TRUE,
  path = tempdir(),
  time_out = 3600,
  job_name,
  verbose = TRUE
)
}
\arguments{
\item{request}{nested list with query parameters following the layout
as specified on the ECMWF APIs page}

\item{user}{user (email address or ID) provided by the ECMWF data service,
used to retrieve the token set by \code{\link[apprs]{apprs_set_key}}}

\item{transfer}{logical, download data TRUE or FALSE (default = TRUE)}

\item{path}{path were to store the downloaded data}

\item{time_out}{how long to wait on a download to start (default =
\code{3*3600} seconds).}

\item{job_name}{optional name to use as an RStudio job and as output variable
name. It has to be a syntactically valid name.}

\item{verbose}{show feedback on processing}
}
\value{
the path of the downloaded (requested file) or the an R6 object
with download/transfer information
}
\description{
Stage a data request, and optionally download the data to disk. Alternatively
you can only stage requests, logging the request URLs to submit download
queries later on using \code{\link[apprs]{apprs_transfer}}.
}
\examples{

\dontrun{
# set key
apprs_set_key(user = "test@mail.com", key = "123")

request <- list(stream = "oper",
   levtype = "sfc",
   param = "167.128",
   dataset = "interim",
   step = "0",
   grid = "0.75/0.75",
   time = "00",
   date = "2014-07-01/to/2014-07-02",
   type = "an",
   class = "ei",
   area = "50/10/51/11",
   format = "netcdf",
   target = "tmp.nc")

# demo query
apprs_request(request = request, user = "test@mail.com")

# Run as an RStudio Job. When finished, will create a
# variable named "test" in your environment with the path to
# the downloaded file.
apprs_request(request = request, user = "test@mail.com", job_name = "test")
}
}
\seealso{
\code{\link[apprs]{apprs_set_key}}
\code{\link[apprs]{apprs_transfer}}
}
\author{
Koen Hufkens
}
