% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apprs_transfer.R
\name{apprs_transfer}
\alias{apprs_transfer}
\title{AppEEARS data transfer function}
\usage{
apprs_transfer(
  url,
  user,
  path = tempdir(),
  filename = tempfile("ecmwfr_", tmpdir = ""),
  verbose = TRUE
)
}
\arguments{
\item{url}{R6 \code{\link[apprs]{apprs_request}}) query output}

\item{user}{user (email address) used to sign up for the ECMWF data service,
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}.}

\item{path}{path were to store the downloaded data}

\item{filename}{filename to use for the downloaded data}

\item{verbose}{show feedback on data transfers}

\item{service}{which service to use, one of \code{webapi}, \code{cds}
or \code{ads} (default = webapi)}
}
\value{
a netCDF of data on disk as specified by a
\code{\link[ecmwfr]{wf_request}}
}
\description{
Returns the contents of the requested url as a NetCDF file downloaded
to disk or the current status of the requested transfer.
}
\examples{

\dontrun{
# set key
apprs_set_key(user = "test@mail.com", key = "123")

# request data and grab url and try a transfer
r <- apprs_request(request, "test@email.com", transfer = FALSE)

# check transfer, will download if available
apprs_transfer(r$get_url(), "test@email.com")
}
}
\seealso{
\code{\link[ecmwfr]{wf_set_key}}
\code{\link[ecmwfr]{wf_request}}
}
\author{
Koen Hufkens
}
