% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_bundle.R
\name{rs_bundle}
\alias{rs_bundle}
\title{AppEEARS list of bundled files to download}
\usage{
rs_bundle(task_id, user)
}
\arguments{
\item{task_id}{task id for which to return the file download bundle}

\item{user}{username used to sign up}
}
\value{
returns a nested list of files to download
}
\description{
Returns a data frame of all data ready for download as one bundle
}
\examples{

\dontrun{
# get a list of datasets
rs_bundle(
user = "your_user_name",
task_id = "a_task_id"
)
}
}
\author{
Koen Hufkens
}
