% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_list_task.R
\name{rs_list_task}
\alias{rs_list_task}
\title{AppEEARS list of tasks and status}
\usage{
rs_list_task(task_id, user)
}
\arguments{
\item{task_id}{task for which to list the status (if missing
all tasks are listed)}

\item{user}{username used to sign up}
}
\value{
returns a data frame with the AppEEARS tasks
}
\description{
Returns a data frame of all submitted tasks either
in full of when providing the di
}
\examples{

\dontrun{
# get a list of datasets
rs_list_task()
}
}
\author{
Koen Hufkens
}
